# services/svc-payment/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Parking - Payment Service API",
        "description": "OpenAPI docs cho svc-payment",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002/api",
            "description": "Local API through svc-auth"
        }
    ],
    "paths": {
...
        "/pricing-rules": {
            "get": {
                "tags": [
                    "Pricing Rules"
                ],
                "summary": "Lấy danh sách bảng giá",
                "description": "Lấy danh sách tất cả bảng giá theo bãi đỗ xe và loại xe",
                "operationId": "4127299580770a6397c9dbd001e14399",
                "parameters": [
                    {
                        "name": "parking_lot_id",
                        "in": "query",
                        "description": "ID bãi đỗ xe",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vehicle_type",
                        "in": "query",
                        "description": "Loại xe (motorbike, car_4_seat, car_7_seat, light_truck)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bảng giá",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "parking_lot_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "vehicle_type": {
                                                "type": "string",
                                                "example": "car_4_seat"
                                            },
                                            "hourly": {
                                                "type": "number",
                                                "example": 10000
                                            },
                                            "daily_cap": {
                                                "type": "number",
                                                "example": 100000
                                            },
                                            "monthly_pass": {
                                                "type": "number",
                                                "example": 600000
                                            },
                                            "peak_enabled": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "peak_multiplier": {
                                                "type": "number",
                                                "example": 1.5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "datetime"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pricing Rules"
                ],
                "summary": "Tạo bảng giá mới",
                "description": "Tạo một bảng giá mới cho bãi đỗ xe và loại xe",
                "operationId": "c3fe3567d99ba84ef6503bee88a27a79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parking_lot_id",
                                    "vehicle_type",
                                    "hourly"
                                ],
                                "properties": {
                                    "parking_lot_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_type": {
                                        "type": "string",
                                        "example": "car_4_seat"
                                    },
                                    "hourly": {
                                        "type": "number",
                                        "example": 10000
                                    },
                                    "daily_cap": {
                                        "type": "number",
                                        "example": 100000
                                    },
                                    "monthly_pass": {
                                        "type": "number",
                                        "example": 600000
                                    },
                                    "peak_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "peak_multiplier": {
                                        "type": "number",
                                        "example": 1.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo bảng giá thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo bảng giá thành công"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/pricing-rules/{id}": {
            "get": {
                "tags": [
                    "Pricing Rules"
                ],
                "summary": "Lấy chi tiết bảng giá",
                "description": "Lấy thông tin chi tiết của một bảng giá",
                "operationId": "c084bea530c4b2281629b1eb31808fcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bảng giá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết bảng giá",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "parking_lot_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "vehicle_type": {
                                                    "type": "string",
                                                    "example": "car_4_seat"
                                                },
                                                "hourly": {
                                                    "type": "number",
                                                    "example": 10000
                                                },
                                                "daily_cap": {
                                                    "type": "number",
                                                    "example": 100000
                                                },
                                                "monthly_pass": {
                                                    "type": "number",
                                                    "example": 600000
                                                },
                                                "peak_enabled": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "peak_multiplier": {
                                                    "type": "number",
                                                    "example": 1.5
                                                },
                                                "parking_lot": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bảng giá"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pricing Rules"
                ],
                "summary": "Cập nhật bảng giá",
                "description": "Cập nhật thông tin của một bảng giá",
                "operationId": "cc1845349009481c5ae9fef3770d7e45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bảng giá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hourly": {
                                        "type": "number",
                                        "example": 12000
                                    },
                                    "daily_cap": {
                                        "type": "number",
                                        "example": 120000
                                    },
                                    "monthly_pass": {
                                        "type": "number",
                                        "example": 720000
                                    },
                                    "peak_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "peak_multiplier": {
                                        "type": "number",
                                        "example": 1.8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật bảng giá thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật bảng giá thành công"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bảng giá"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pricing Rules"
                ],
                "summary": "Xóa bảng giá",
                "description": "Xóa một bảng giá",
                "operationId": "813ee145aaba3342109d5ad3eb917b00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bảng giá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa bảng giá thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa bảng giá thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bảng giá"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}